snapcraft: fix build using golang 1.13+
golang 1.13 and 1.14 will both fail with different errors when
attempting `go get` commands.  Since full support of go.mod is
not yet available (it is available in snapcraft v3.10), this commit
adds GO111MODULE=off to the build environment.

This also locks golang into version 1.13 using the go plugin's
`go-channel` property, which defaults to the latest `stable`,
which is currently 1.13.  Using this property will allow falling
back to 1.12 without the requirement to GO111MODULE=auto.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>